+++ src/engine_arrays.c
@@ -38,7 +38,7 @@ short iso_td[TD_ISO_POINTS];
 unsigned short floor_to_ceiling_map[TEXTURE_BLOCKS_COUNT];
 struct WibbleTable blank_wibble_table[128];
 
-long randomisors[512];
+int32_t randomisors[512];
 struct WibbleTable wibble_table[128];
 long floor_height_table[256];
 long lintel_top_height[256];
