+++ src/lua_api.c
@@ -755,8 +755,8 @@ static int lua_Reveal_map_location(lua_State *L)
     SYNCDBG(0, "Revealing location type %lu", target);
     for (PlayerNumber i = player_range.start_idx; i < player_range.end_idx; i++)
     {
-        long x = 0;
-        long y = 0;
+        int32_t x = 0;
+        int32_t y = 0;
         find_map_location_coords(target, &x, &y, i, __func__);
         if ((x == 0) && (y == 0))
         {
@@ -1960,7 +1960,7 @@ static int lua_get_rooms_of_player_and_kind(lua_State *L)
     const char* room_name = luaL_checkstring(L, 2);
 
     lua_newtable(L);
-    unsigned long k = 0;
+    uint32_t k = 0;
 
     for (PlayerNumber i = player_range.start_idx; i < player_range.end_idx; i++)
     {
